+++ pkg/queryfrontend/roundtrip.go
@@ -70,16 +70,13 @@ func NewTripperware(config Config, reg prometheus.Registerer, logger log.Logger)
 	if err != nil {
 		return nil, err
 	}
-	queryInstantTripperware, err := newInstantQueryTripperware(
+	queryInstantTripperware := newInstantQueryTripperware(
 		config.NumShards,
 		queryRangeLimits,
 		queryInstantCodec,
 		prometheus.WrapRegistererWith(prometheus.Labels{"tripperware": "query_instant"}, reg),
 		config.ForwardHeaders,
 	)
-	if err != nil {
-		return nil, err
-	}
 	return func(next http.RoundTripper) http.RoundTripper {
 		return newRoundTripper(next, queryRangeTripperware(next), labelsTripperware(next), queryInstantTripperware(next), reg)
 	}, nil
@@ -191,10 +188,7 @@ func newQueryRangeTripperware(
 	}
 
 	if numShards > 0 {
-		analyzer, err := querysharding.NewQueryAnalyzer()
-		if err != nil {
-			return nil, errors.Wrap(err, "create query analyzer")
-		}
+		analyzer := querysharding.NewQueryAnalyzer()
 		queryRangeMiddleware = append(
 			queryRangeMiddleware,
 			PromQLShardingMiddleware(analyzer, numShards, limits, codec, reg),
@@ -332,14 +326,11 @@ func newInstantQueryTripperware(
 	codec queryrange.Codec,
 	reg prometheus.Registerer,
 	forwardHeaders []string,
-) (queryrange.Tripperware, error) {
+) queryrange.Tripperware {
 	instantQueryMiddlewares := []queryrange.Middleware{}
 	m := queryrange.NewInstrumentMiddlewareMetrics(reg)
 	if numShards > 0 {
-		analyzer, err := querysharding.NewQueryAnalyzer()
-		if err != nil {
-			return nil, errors.Wrap(err, "create query analyzer")
-		}
+		analyzer := querysharding.NewQueryAnalyzer()
 		instantQueryMiddlewares = append(
 			instantQueryMiddlewares,
 			queryrange.InstrumentMiddleware("sharding", m),
@@ -352,7 +343,7 @@ func newInstantQueryTripperware(
 		return queryrange.RoundTripFunc(func(r *http.Request) (*http.Response, error) {
 			return rt.RoundTrip(r)
 		})
-	}, nil
+	}
 }
 
 // shouldCache controls what kind of Thanos request should be cached.