+++ pkg/querysharding/analyzer_test.go
@@ -120,7 +120,7 @@ sum by (container) (
 		{
 			name:           "multiple binary expressions with grouping",
 			expression:     `(http_requests_total{code="400"} + on (pod) http_requests_total{code="500"}) / on (cluster, pod) http_requests_total`,
-			shardingLabels: []string{"cluster", "pod"},
+			shardingLabels: []string{"pod"},
 		},
 		{
 			name:           "histogram quantile",
@@ -136,6 +136,11 @@ sum by (container) (
 			name:           "subquery with function",
 			expression:     "increase(sum(http_requests_total) by (pod, cluster) [1h:1m])",
 			shardingLabels: []string{"cluster", "pod"},
+		},
+		{
+			name:           "ignore vector matching with 2 aggregations",
+			expression:     `sum(rate(node_cpu_seconds_total[3h])) by (cluster_id, mode) / ignoring(mode) group_left sum(rate(node_cpu_seconds_total[3h])) by (cluster_id)`,
+			shardingLabels: []string{"cluster_id"},
 		},
 	}
 
@@ -177,8 +182,7 @@ http_requests_total`,
 
 	for _, test := range nonShardable {
 		t.Run(test.name, func(t *testing.T) {
-			analyzer, err := NewQueryAnalyzer()
-			require.NoError(t, err)
+			analyzer := NewQueryAnalyzer()
 			analysis, err := analyzer.Analyze(test.expression)
 			require.NoError(t, err)
 			require.False(t, analysis.IsShardable())
@@ -187,8 +191,7 @@ http_requests_total`,
 
 	for _, test := range shardableByLabels {
 		t.Run(test.name, func(t *testing.T) {
-			analyzer, err := NewQueryAnalyzer()
-			require.NoError(t, err)
+			analyzer := NewQueryAnalyzer()
 			analysis, err := analyzer.Analyze(test.expression)
 			require.NoError(t, err)
 			require.True(t, analysis.IsShardable())
@@ -202,8 +205,7 @@ http_requests_total`,
 
 	for _, test := range shardableWithoutLabels {
 		t.Run(test.name, func(t *testing.T) {
-			analyzer, err := NewQueryAnalyzer()
-			require.NoError(t, err)
+			analyzer := NewQueryAnalyzer()
 			analysis, err := analyzer.Analyze(test.expression)
 			require.NoError(t, err)
 			require.True(t, analysis.IsShardable())